Safari-609.1.11.1/LayoutTests/fast/events/before-input-delete-empty-list-target-ranges-expected.txt

To manually test this, delete the empty list element and observe the output.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.getTargetRanges().length is 1
PASS range.startContainer.nodeName is '#text'
PASS range.startContainer.textContent is 'foo'
PASS range.endContainer is item
PASS range.startOffset is 3
PASS range.endOffset is 0
PASS successfullyParsed is true

TEST COMPLETE
foo
